+++ packages/vendure-plugin-accept-blue/src/api/accept-blue-handler.ts
@@ -91,13 +91,42 @@ export const acceptBluePaymentHandler = new PaymentMethodHandler({
       type: 'boolean',
       required: false,
       defaultValue: true,
-      label: [{ languageCode: LanguageCode.en, value: 'Allow Pay' }],
+      label: [{ languageCode: LanguageCode.en, value: 'Apple Pay' }],
     },
     tokenizationSourceKey: {
       type: 'string',
       required: false,
       label: [
         { languageCode: LanguageCode.en, value: 'Hosted tokenization key' },
+      ],
+    },
+    googlePayMerchantId: {
+      type: 'string',
+      required: false,
+      label: [
+        { languageCode: LanguageCode.en, value: 'Google Pay Merchant Id' },
+      ],
+      description: [
+        {
+          languageCode: LanguageCode.en,
+          value: 'The Merchant Id as defined in your Google Payment Profile',
+        },
+      ],
+    },
+    googlePayGatewayMerchantId: {
+      type: 'string',
+      required: false,
+      label: [
+        {
+          languageCode: LanguageCode.en,
+          value: 'Google Pay Gateway Merchant Id',
+        },
+      ],
+      description: [
+        {
+          languageCode: LanguageCode.en,
+          value: 'Can be found in your Accept Blue Control Panel',
+        },
       ],
     },
     webhookSecret: {