+++ packages/vendure-scripts/src/index.ts
@@ -2,3 +2,4 @@ export * from './assign-products-to-channel';
 export * from './assign-customers-to-channel';
 export * from './assign-orders-to-channel';
 export * from './copy-database';
+export * from './populate-eu-zones-and-tax-rates';